% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellsolve.R
\name{wellsolve}
\alias{wellsolve}
\title{Minimize the Total Pumping Cost}
\usage{
wellsolve(
  projected,
  T,
  R,
  ff,
  qtot,
  r0,
  Dlist,
  qextra,
  nw,
  xc,
  yc,
  links,
  nwextra,
  xextra,
  yextra,
  xstart,
  g = 9.81
)
}
\arguments{
\item{projected}{A character string of projection arguments;
the arguments must be entered exactly as in the PROJ.4 documentation;
if the projection is unknown, use as.character(NA),
it may be missing or an empty string of zero length and
will then set to the missing value.
With rgdal built with PROJ >= 6 and GDAL >= 3,
the +init=key may only be used with value epsg:<code>.}

\item{T}{Aquifer’s transmissivity in square meters per second.}

\item{R}{Radius of influence of the system of wells in meters.}

\item{ff}{Friction coefficient along the pipes}

\item{qtot}{Total required flow rate in qubic meters per second.}

\item{r0}{Diameter of all wells in meters.}

\item{Dlist}{A vector with the diameters of the pipes in meters.}

\item{qextra}{A vector with the required flow rates for the extra wells
in qubic meters per second.}

\item{nw}{Number of wells + 1 (the tank).}

\item{xc}{A vector with x coordinates. The first value is the tank and
the rest are the wells.}

\item{yc}{A vector with y coordinates. The first value is the tank and
the rest are the wells.}

\item{links}{A two-column matrix describing the links between the wells. Each
row describes one link. The two columns contain the id numbers of the
connected wells. The id number of the tank is 0.}

\item{nwextra}{Number of extra wells.}

\item{xextra}{A vector with x coordinates for the extra wells.}

\item{yextra}{A vector with y coordinates for the extra wells.}

\item{xstart}{A wild guess about the result.
This helps the algorithm search for the result in the correct area}

\item{g}{Acceleration of gravity.}
}
\value{
A list containing the following:
\item{qresult}{The optimal well flows}
\item{hf}{The hydraulic head losses}
\item{s}{The hydraulic head level drawdowns}
\item{printmap}{The diagram of the tank and the wells}
\item{sfinal}{The hydraulic head level drawdowns including
those of the extra wells}
\item{kx}{The xi constant of the Darcy-Weisbach formula}
\item{QQsqure}{The square of the pipe flow rates}
}
\description{
Minimize the Total Pumping Cost from an Aquifer
to a Water Tank, Via a Pipe Network
}
\examples{
library(wellsolve)

projected <- "+proj=utm +zone=34"
x0 = 650000
y0 = 4400000

# the first value is the tank and the rest are wells
xc <- c() ; xc <- c(0,100,180,100,700,800,900) ; xc <- xc + x0 -10000
yc <- c() ; yc <- c(0,0,0,80,0,0,900) ; yc <- yc + y0 -10000

# xstart is a wild guess about the result.
# This helps the algorithm search for the result in the correct area
xstart<- c(0.250,0.1,0.25,0.1,0.1,0.01)

nw=7 #number of wells + 1 (the tank)
links<- data.frame(matrix(ncol=2, nrow=(nw-1)))
links[1:(nw-1),1]<-1:(nw-1)
links[1,2]<-0 ; links[2,2]<-1 ; links[3,2]<-0
links[4,2]<-2 ; links[5,2]<-4 ; links[6,2]<-2

Dlist<-list()
Dlist[1:6]<-0.05


# number of extra wells
nwextra <- 2
xextra<- c() ; xextra <- c(500,800) ; xextra <- xextra + x0 -10000
yextra<- c() ; yextra <- c(500,300) ; yextra <- yextra + y0 -10000
qextra<- c() ; qextra[1]<- 0; qextra[2]<- 0

result<-wellsolve(projected, T=0.0025 ,R=3000, ff=0.03, qtot<-0.5, r0=0.200,
                  Dlist,qextra,nw,xc,yc,links,nwextra,xextra,yextra, xstart)

# Case A
Dlist[1:6]<-2
result<-wellsolve(projected, T=0.0025 ,R=3000, ff=0.03, qtot<-0.5, r0=0.200,
                  Dlist,qextra,nw,xc,yc,links,nwextra,xextra,yextra, xstart)

# Case LC1
result<-wellsolve(projected, T=0.0025 ,R=3000, ff=0, qtot<-0.5, r0=0.200,
                  Dlist,qextra,nw,xc,yc,links,nwextra,xextra,yextra, xstart)

# Case B
Dlist[1:6]<-0.4
result<-wellsolve(projected, T=0.0025 ,R=3000, ff=0.03, qtot<-0.5, r0=0.200,
                  Dlist,qextra,nw,xc,yc,links,nwextra,xextra,yextra, xstart)

# Case C
qextra[1] <- 0.1
qextra[2] <- 0.4
result<-wellsolve(projected, T=0.0025 ,R=3000, ff=0.03, qtot<-0.5, r0=0.200,
                  Dlist,qextra,nw,xc,yc,links,nwextra,xextra,yextra, xstart)
}
\references{
Nagkoulis, N., Katsifarakis, K. Minimization of Total Pumping Cost from
an Aquifer to a Water Tank, Via a Pipe Network.
Water Resour Manage 34, 4147–4162 (2020).
\doi{10.1007/s11269-020-02661-x}
}
